requirements.txt
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.50.0
playwright>=1.40.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
typer>=0.12.0
loguru>=0.7.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
twilio>=9.0.0
flask>=3.0.0
